chore: move files to search package
