+++ libraries/cfg/tsconfig.json
@@ -5,5 +5,5 @@
     "rootDir": ".",
     "composite": true
   },
-  "include": ["**/*.ts"]
+  "include": ["**/*.ts", "../search/lib/util/ControlFlowPath.ts"]
 }
